fix critical and medium-severity issues across pyridecore
- sprintf -> snprintf in DeviceController.cpp and VideoToWebBridge.cpp
- fix iterator deref before end() check in ServerDataProcessor.cpp
- add NULL checks in AppConfigManager.cpp (enablePySvrElem, nameElem, etc.)
- check opus_decoder_create return code in AudioDataReceiver.cpp
- fix sign-extension bug in PyRideNetComm.cpp (char -> unsigned char)
- fix socket FD leak on partial init in DeviceController.cpp
- move EVP_CIPHER_CTX_init inside OpenSSL < 1.1 guard only
- TOCTOU race: read isTimerExecuting/isTimerRunning under lock
- add timeout guard in delTimer busy-loop
- use unsigned short for dataCount, fix underflow check
- guard telemetryRobot_-- against underflow
- fix RTP seq num wrap (uint16_t arithmetic)
- inline constexpr for per-TU duplicated arrays
- guard secureSHA256Hash with USE_ENCRYPTION
- add stdio.h before jpeglib.h for FILE decl

+++ RTPDataReceiver.cpp
@@ -93,7 +93,7 @@ int RTPDataReceiver::grabData(unsigned char **dataBuffer,
       dataBuffer_ = new unsigned char[dataBufferSize_];
     }
     // copy data
-    int seqNum = adu->getSeqNum();
+    uint16_t seqNum = adu->getSeqNum();
     if (seqNum - lastSeqNum_ > 10) {
       printf("RTPReceiver: loss of more than 10 packets\n");
     }
